Moved function promptWhenGstUnavailable to gst.js
